requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
albumentations=1.3.0=pypi_0
blas=1.0=openblas
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
cairo=1.16.0=hb05425b_3
certifi=2022.12.7=pyhd8ed1ab_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=unix_pyhd8ed1ab_2
cloudpickle=2.2.1=pypi_0
contourpy=1.0.5=py310hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
eigen=3.3.7=hd09550d_1
expat=2.4.9=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_1
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=4.3.0=hd55b92a_0
hdf5=1.10.6=h3ffc7dd_1
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imageio=2.26.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lazy-loader=0.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libopenblas=0.3.21=h043d6bf_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvmlite=0.39.1=pypi_0
loguru=0.6.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.6.2=py310h06a4308_0
matplotlib-base=3.6.2=py310h945d387_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numba=0.56.4=pypi_0
numexpr=2.8.4=py310h757a811_0
numpy=1.23.5=py310hac523dd_0
numpy-base=1.23.5=py310h375b286_0
opencv=4.6.0=py310hefb4dc4_2
opencv-python-headless=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1s=h7f8727e_0
packaging=22.0=py310h06a4308_0
pandas=1.5.2=py310h1128e8f_0
pcre=8.45=h295c915_0
pillow=9.3.0=py310hace64e9_1
pip=22.3.1=py310h06a4308_0
pixman=0.40.0=h7f8727e_1
ply=3.11=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
python=3.10.9=h7a1cb2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.1=pyhd8ed1ab_0
pytz=2022.7=py310h06a4308_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.0=py310h6a678d5_0
scipy=1.9.3=py310heeff2f4_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py310h06a4308_0
shap=0.41.0=pypi_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
slicer=0.0.7=pypi_0
sqlite=3.40.1=h5082296_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.3.15=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torch=1.13.1+cu116=pypi_0
torchaudio=0.13.1+cu116=pypi_0
torchvision=0.14.1+cu116=pypi_0
tornado=6.2=py310h5eee18b_0
tqdm=4.65.0=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
